Mech 2
I'm building a casing out of 2 Altoid contrainers, soft velcro padding, and using potting to water proof and secure the board in a modular form. I will supply an power supply of 6xLR41 batteries (less readily accessible but longer lasting), which can power the system when outside of the case, as well as a converter to a 9V power supply (more readily accessible but shorter life).